rename WikiCreateEventListener
